add ElasticsearchConfig
